refactor: moveRule 패키지명 수정 및  updateTurn과 updateFinished를 update로 통합
