+++ src/main/java/janggi/domain/piece/Elephant.java
@@ -1,7 +1,7 @@
 package janggi.domain.piece;
 
-import janggi.domain.mouveRule.ElephantMoveRule;
-import janggi.domain.mouveRule.MoveRule;
+import janggi.domain.moveRule.ElephantMoveRule;
+import janggi.domain.moveRule.MoveRule;
 
 public class Elephant extends Piece {
