+++ src/main/java/janggi/domain/piece/Tank.java
@@ -1,7 +1,7 @@
 package janggi.domain.piece;
 
-import janggi.domain.mouveRule.MoveRule;
-import janggi.domain.mouveRule.TankMoveRule;
+import janggi.domain.moveRule.MoveRule;
+import janggi.domain.moveRule.TankMoveRule;
 
 public class Tank extends Piece {
